# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ErrorInfo:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/dto.ValidationDetail'
        type: array
      help:
        type: string
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: string
    type: object
  dto.Meta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ErrorInfo'
      meta:
        $ref: '#/definitions/dto.Meta'
      success:
        type: boolean
    type: object
...
  handler.WarehouseListResponse:
    description: Warehouse list item with basic information
    properties:
      city:
        example: Guangzhou
        type: string
      code:
        example: WH-001
        type: string
      created_at:
        example: "2026-01-24T12:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_default:
        example: true
        type: boolean
      name:
        example: Main Warehouse
        type: string
      province:
        example: Guangdong
        type: string
      short_name:
        example: Main WH
        type: string
      sort_order:
        example: 0
        type: integer
      status:
        enum:
        - enabled
        - disabled
        example: enabled
        type: string
      type:
        enum:
        - normal
        - virtual
        - transit
        example: normal
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@erp.example.com
    name: API Support
    url: https://github.com/erp/backend
  description: 进销存系统后端 API - 基于 DDD 设计的库存管理系统
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ERP Backend API
  version: "1.0"
paths:
...
  /partner/warehouses:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of warehouses with optional filtering
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Search term (name, code)
        in: query
        name: search
        type: string
      - description: Warehouse status
        enum:
        - active
        - inactive
        in: query
        name: status
        type: string
      - description: Warehouse type
        enum:
        - physical
        - virtual
        - consign
        - transit
        in: query
        name: type
        type: string
      - description: City
        in: query
        name: city
        type: string
      - description: Province
        in: query
        name: province
        type: string
      - description: Filter by default status
        in: query
        name: is_default
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        maximum: 100
        name: page_size
        type: integer
      - default: is_default
        description: Order by field
        in: query
        name: order_by
        type: string
      - default: desc
        description: Order direction
        enum:
        - asc
        - desc
        in: query
        name: order_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.WarehouseListResponse'
                  type: array
                meta:
                  $ref: '#/definitions/dto.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: List warehouses
      tags:
      - warehouses
    post:
      consumes:
      - application/json
      description: Create a new warehouse in the partner module
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Warehouse creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWarehouseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.WarehouseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new warehouse
      tags:
      - warehouses
  /partner/warehouses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a warehouse by ID (cannot delete default warehouse)
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Warehouse ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Delete a warehouse
      tags:
      - warehouses
    get:
      consumes:
      - application/json
      description: Retrieve a warehouse by its ID
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Warehouse ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.WarehouseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Get warehouse by ID
      tags:
      - warehouses
    put:
      consumes:
      - application/json
      description: Update an existing warehouse's details
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Warehouse ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Warehouse update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateWarehouseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.WarehouseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Update a warehouse
      tags:
      - warehouses
...
  /partner/warehouses/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable an active warehouse
      parameters:
      - description: Tenant ID (optional for dev)
        in: header
        name: X-Tenant-ID
        type: string
      - description: Warehouse ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.WarehouseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorInfo'
              type: object
      security:
      - BearerAuth: []
      summary: Disable a warehouse
      tags:
      - warehouses
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"